ponderada_8/requirements.txt
openai<1
python-dotenv
PyPDF2
faiss-cpu
langchain
chainlit
requests
tiktoken